Correção: o desafio 3 não havia sido feito

+++ 02 - Desafios GitHub/Básico/numeros 2.py
@@ -2,5 +2,5 @@
 # e que apresente o número informado com duas casas decimais
 # pra ficar mais divertido use float, pra receber qualquer número.
 
-numero = int(input('Informe um número inteiro qualquer: '))
+numero = float(input('Informe um número inteiro qualquer: '))
 print(f'Este número com duas casas decimais fica assim: {numero:.2f}')
